Move C&C world tooltip code into ViewportControllerWidget.
