+++ OpenRA.Mods.Cnc/OpenRA.Mods.Cnc.csproj
@@ -90,7 +90,6 @@
     <Compile Include="SpawnViceroid.cs" />
     <Compile Include="TiberiumRefinery.cs" />
     <Compile Include="Widgets\CncWidgetUtils.cs" />
-    <Compile Include="Widgets\CncWorldInteractionControllerWidget.cs" />
     <Compile Include="Widgets\Logic\ButtonTooltipLogic.cs" />
     <Compile Include="Widgets\Logic\CncConquestObjectivesLogic.cs" />
     <Compile Include="Widgets\Logic\CncIngameChromeLogic.cs" />
